+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.0] - 4/11/2025
+
+### Added
+
+ - Added atom for email input
+ - Validation input function ValidateEmail created
+
 ## [0.1.0] - 3/11/2025
 
 ### Added